Mostrar mensajes de confirmación y espera

+++ 07-journal/src/modules/daybook/views/EntryView.vue
@@ -49,6 +49,7 @@
 <script>
 import { defineAsyncComponent } from 'vue'
 import { mapGetters, mapActions } from 'vuex'
+import Swal from 'sweetalert2'
 
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
@@ -90,6 +91,12 @@ export default {
             this.entry = entry
         },
         async saveEntry() {
+
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
             
             if ( this.entry.id ) {
                 // Actualizar
@@ -104,14 +111,33 @@ export default {
                 // await action 
 
                 // redirectTo => entry, params: id
-            }    
+            }  
+            
+            Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
         },
 
         async onDeleteEntry() {
 
-            await this.deleteEntry( this.entry.id )
+            const { isConfirmed } = await Swal.fire({
+                title: '¿Está Seguro?',
+                text: 'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
+
+            if ( isConfirmed ) {
+                new Swal({
+                    title: 'Espere por favor',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                await this.deleteEntry( this.entry.id )
+                this.$router.push({ name: 'no-entry' })
+
+                Swal.fire('Eliminado', '', 'success')
+            }
+
             // console.log('delete', this.entry );
-            this.$router.push({ name: 'no-entry' })
 
         }
     },
